requirements.txt
# Core AI Framework
crewai>=0.148.0
crewai-tools>=0.55.0

# Web Interface
panel>=1.7.4

# Data & Validation  
pydantic>=2.11.0
pandas>=2.3.0

# AI & LLM
openai>=1.97.0
langchain>=0.3.26
langchain-openai>=0.2.14
langchain-community>=0.3.27

# Web Deployment
python-dotenv>=1.1.0
uvicorn>=0.35.0

# Utilities
markdown>=3.8.0
requests>=2.32.0

